Make try_int() raise rather than return None (#124)
Fail faster when we don't want to allow `None`. Rename existing impl to
`try_int_or_none`.

General principle here: simple validation functions where a single input
is returned can be in four general categories based on nullability:
```
1. input is non-nullable, output is non-nullable
2. input is nullable, output is non-nullable
3. input is nullable, output is nullable
4. input is non-nullable, output is nullable
```

IMO, that's also their rank from most to least useful.

For (1), mypy (or whatever your language type linter is) can flag
potential nulls passed in, and we can guarantee a non-null value out.

For (2), we're saying that it's never acceptable for an input to be
nullable but that incomplete type information somewhere downstream
(e.g., from a map.get()) can't guarantee a null input based on type, so
we have to check at runtime.

For (3) & (4), the validation contract may be less clear. Is it ok for
the input to be null, or not? And even if the contract is clear, from a
type-checking perspective we might be passing in a value that our
type-checking system is confident isn't null and returning a value (the
same value!) that it now thinks may be null.

I think the contracts in (1) and (3) are the clearest and think it's
worth having both function signatures if use cases call for them.