Simplify vi_cell_blu_parser (#144)
We shouldn't be using `iterrows`. Not just because it's bad for perf
(e.g.,
https://towardsdatascience.com/efficiently-iterating-over-rows-in-a-pandas-dataframe-7dd5f9992c01)
but because it erases type information by returning a `Series` of values
of different types.

This cleans up usage of `iterrows` in vi-cell-blu. Test util changes are
needed to consider `int` and `np.int64` the same for tests since types
are now preserved better in src.

---------

Signed-off-by: Brian Worth <brian@benchling.com>