Create NamedFileContents NamedTuple (#162)
File contents and name are closely linked, so create a NamedTuple to
pass them around together.

Bonus: getting rid of unused parameter warnings for VendorParser
children that don't use the file name.

Out of scope: changing signatures of methods in `to_allotrope`. That may
happen later but should be accompanied by changes to the README and
changelog as they'd break backwards compatibility.