chore: Upgrade datamodel-codegen and regenerate schema models (#244)
In preparation for a major update to schema generation script, we need
to upgrade datamodel-codegen package.

This PR upgrades to the latest version, and updates all generated models
based on the new version. This will reduce diffs in the real change PR.

All changes are just classes moving around in the file, no classes were
changed/added/deleted.