Improve unit and value dataclasses (#243)
Units shouldn't be optional, and mypy should know that. Eliminate
unnecessary classes for `Optional` units and make a `HasUnit` parent for
unit dataclasses. No functional changes, just better types.

Note that this change triggers warnings in PyCharm in
`../shared/definitions/custom/py`. This appears to be a manifestation of
https://youtrack.jetbrains.com/issue/PY-60352/False-positive-Inherited-non-default-arguments-warning-with-multiple-inheritances.
Setting a default value for `unit` in `HasUnit` fixes this warning, but
we don't want a default value.

---------

Signed-off-by: Brian Worth <brian@benchling.com>