Test Model in test_model_from_file_absorbance() (#283)
Make this a better unit test by testing (part of) the in-memory `Model`.
The exact serialization of the `Model` is tested in
`test_to_allotrope_absorbance()`.

Note that this is the only unit test for `allotrope_model_from_file`. It
is called in src by `allotrope_from_file`, which has lots of unit tests.