Simplify vi_cell_blu_reader (#150)
Simply calling `pd.read_csv()` is sufficient.

Downstream in `vi_cell_blu_parser`, we request specific types, and
ultimately all we're doing with these values is to write them to JSON.
So we don't really need to be converting strings to numbers and back to
strings. Parsing date strings to canonical dates and then back to ISO
also happens downstream, specifically in
`VendorParser._get_date_time()`.