feat: Improve schema model generation script (#205)
This PR makes major improvements to the schema generation library, which
is responsible for translating JSON schemas from allotrope into python
models for use in parser code. The improvements make it possible for the
library to handle the more modular schemas that are being created in
recent months, and pave the way for the ability to read the schemas
directly from the online source and for modularizing the generated
python classes themselves.

The primary changes in this PR are:

1. Adds `schema_cleaner`, which loads a schema before it is passed to
the python dataclass generation tool (`datamodel_codegen`) to fix
various issues that prevent the tool successfully processing some
schemas.

Most of these issues come down to the use of compound schemas that use
"allOf", "oneOf", or "anyOf" operators to compose schemas. Examples of
things the tool can not handle include: allOf values with one value,
anyOf values nested within allOf values, combing schemas in an allOf
that only specify required fields, etc.

The schema cleaner also automatically adjusts hyperlink schema
references to point to local copies of these schemas, another step that
used to have to be done manually.

Finally, the schema cleaner tracks all units referenced in all schemas,
and automatically generated the units schema/python classes, as well as
the `tQuantityValue` items corresponding to those units.

2. Makes updates to `model_class_editor` to further fix issues that are
present after the dataclass generation completes.

A particular focus here is wrangling the combinatorial explosion that
can happen due to uses of several compound schemas. e.g. a pair of anyOf
combinations can create a cross product of dozens of classes. Though it
is technically correct to have all of these classes, it makes the
library basically unusable.

Instead, we combine these classes, by allowing the fields that conflict
to be a union of possible fields, or by combining two different versions
of a dataclass that have many shared fields and then one extra optional
field each into a single dataclass with all options (note that we do not
do this if there are conflicting required keys!).

The result of these changes is not 100% faithful to the schema, at it
could allow a user to set two fields on a dataclass that are mutually
exclusive in the json schema. However, it is the only way to make these
classes usable for writing parsers. Future work will endeavor to add
logic to the combined dataclasses to enforce the exclusion logic.

3. This PR also adds `backup_manager`, which provides a context manager
to automatically backup and restore files during schema generation, so
that an error does not leave the files in a bad state.


All schemas are regenerated with the changes above. Most of the existing
schemas are nearly identical, with only some minor changes related to
automatic alphabetization of dataclass fields / enforcement that
required fields are placed before optional fields. A couple existing
schemas also have some duplicated classes simplified away, but nothing
that materially affects the result.

Additionally, the `liquid-chromatography` `2023/09` version is added.
This schema was previously impossible to add due the structure of the
schema. This demonstrates that the new, modularized schemas that are
coming from allotrope recently can be handled by the schema generation
library going forward.