feat: Add structuring logic to allotropy and refactor into converter (#429)
We use `cattrs` to "unstructure" the python dataclasses into JSON

This PR adds logic to use `cattrs` to "structure" ASM JSON back into
python classes.

This can be used by users of allotropy to load Allotropy output into
python in order to process the data.

The benefit of bringing this into allotropy is that handle special cases
(e.g. SPECIAL_KEYS) when structuring for the user, and we can test that
all allotropy output (test cases, at least) can be structured.