test: Allow test util to write new test output and improve error messages (#487)
This PR makes a couple upgrades to test_util:

- Allow writing a missing test file if
`write_actual_to_expected_on_fail=True`. The standard way we add new
test cases it to turn on `write_actual_to_expected_on_fail` and run to
generate the test output. But, before you had to make a dummy file or
the test util would fail due to the missing file. Fix that.
- Improve test util error messages:
    - When overwriting file, write a helpful message
- Use DeepDiff.pretty() to write human readable messages for deepdiff
    - Remove optional verbose message now that error is more readable.