refactor: Remove converting measurement calc docs to global calc docs (#558)
This removes a mistaken approach of pulling all calculated documents
from all measurement documents into the global calculated data documents
list. This is wrong because some ASM documents report calculated data
documents inside of measurement documents.

The location of the calculated data documents within the data classes
should signal where they should be placed within the ASM document.

This was done in the `plate reader` mapper because one parser (unchained
labs) was taking this approach, of putting all calculated data documents
in the global document, even if they correspond to measurement rows. I'm
not sure why we are putting measurement-specific calculated data
documents in the global calculated data document, and I'm not changing
it here (left a TODO comment to investigate).

Instead, the offending parser is changed to move the calculated
documents to the global level in the data class.

Note that this is classified as a refactor only, because it does not
change any output of any test cases.