fix: Fix measurement grouping in roche cedex bioht (#567)
This fixes a rare(ish?) edge case in bioht where two measurements for
the same sample and property are taken over several days.

In the result files for this parser, all measurements are taken at
separate times, but we want to group measurements for a given sample. To
do this, all measurements for a sample were group, and then duplicate
measurements were aggregated in a list and split out in to separate
measurement documents in the order the were encountered.

However, this misses the context of what these duplicate measurements
actually are. In reality, a duplicate measurement for a sample happens
when multiple measurements are taken over a period of time (in the
example file we have, over 3 days).

To really capture this context, we instead group measurements close in
time (using a heuristic of 1 hour max time difference)

Note that this also fixes another minor bug for the recorded measurement
time, which the requirements specify should be the earliest measurement
time (previous it was just the first measurement time encountered).

---------

Signed-off-by: Nathan Stender <nathan.stender@benchling.com>