fix: improve identification of numeric columns for sofmax pro calculated data (#588)
Some columns in the SoftMax Pro adapter were incorrectly identified as
not numeric for calculated data documents. This was due to some values
being flagged as `Masked` or `Range?`. Those cases will be treated as
`NaN`.

Note: in a future version of this adapter, those values will probably be
reported in the `ErrrorDocument` when we drop support for
`InvalidJsonFloat` in the `tQuantityValues`

---------

Co-authored-by: Nathan Stender <nathan.stender@benchling.com>