chore: Add new actions to git pre-push-check (#611)
This PR improves our pre-push hook by making it easier to add pre-push
hooks, and by adding two new actions:

1. Checks for uncommitted changes and blocks if any are found
2. Runs short tests (`hatch run test -m "not long"`) and blocks if tests
fail

As part of this change - remove `short` tag and instead use `long` tag.
This is done because it is much easier to track and mark long tests than
it is to track every short test (which is why most short tests were not
marked). Especially because we can mark all parsing tests as `long` via
the base test.