refactor: Use schema mapper for BMG Mars parser (#620)
As part of this change, `measurement_time`, `analyst`, and
`experiment_type` are moved to `MeasurementGroup`

This brings the dataclass fields more in line with the organization of
the ASM schema structure.

This change required modifying the other parsers that use the
plate_reader schema mapper in trivial ways (moving fields around). In
one case, `CtlImmunospot` I ended up changing a method from directly
creating `Metadata` to creating a `Header` dataclass that is then used
to create `Metadata` and `MeasurementGroups`

Finally, note the only test data changes was that the measurement
identifiers for the BMG Mars parser changed order. This makes no
different in production, as the IDs are random uuids, and even in test
makes them make more sense (they are now sequential in the test file)