refactor: Change the way custom information is organized in schema mappers and ASM outputs to be consistent with future expectations of ASM (#673)
The new policies for custom information documents in ASM are:

1. If we expect the data to be added to ASM in the future: add the data
directly to the schema mapper dataclass, and then add it to the `custom
information document` in the schema mapper, with a comment #TODO(ASM
gaps)`. In these cases, the `custom information document` is a stop-gap
until ASM catches up.
2. If the data is device/customer specific and should not be in ASM: add
to a `custom_info` dict of the schema mapper dataclass, and pass through
in the schema mapper. In this case `custom information document` is a
feature - allowing the ASM output to provide extra data that the user
might need. This conforms with new ASM schemas coming out now, which
have `custom information document`s embedded.

The only change in test data is that `omit` in the AppBio Quantstudio
custom data moves from measurement document custom info to processed
data custom info. This is because `omit` is in the `[Results]` section
of the raw data, which corresponds to processed data.