feat: Thermo Fisher Scientific Genesys On-Board - Initial implementation (#792)
Add adapter for the Thermo Fisher Scientific Genesys On-Board software.
Since this adapter requirements are almost identical to the
`VISIONLIte`, the parser just calls the `VISIONLite` reader and data
creator for now. This might change in the future if new example files
make the requirements diverge enough that it is best to have two
separate parsers for better maintainability.