fix: Beckman Echo Plate Reformat - update measurement time to fill missing date with file date if needed (#990)
One example file has the per-measurement time missing a date (it
provides time only).

In this case, the default behavior of the dateutil library is to use
today's date, which is incorrect.

Instead, fill the date with the file header's date, is possible.

If the measurement time is missing a date, and the header has no date,
throw an error.