feat: Molecular Devices SoftMax Pro - update logic on group data to get calculated entries and custom information (#992)
Previously, the columns used as calculated data from the group block
table were identified by analyzing the first row of each Sample in the
table and checking whether the value was a number or in a list of
expected "numeric errors".

This PR adds a classification of the group data columns by checking if
any of the values in the columns can be considered numeric before
grouping by sample. This allows us to classify anything else in the
group block table as "custom information". This custom information is
latter added to the measurement documents.

Co-authored-by: Nathan Stender <nathan.stender@benchling.com>