feat: Molecular Devices SoftMax Pro - Enable Reduced-only file parsing (#1016)
Updating the Softmax parser to pull in the following data:

    Group: Control
    Sample	Wells	Sample#	Values	Results	MeanResults	CVs	CVresults	
    
    Group: Matrix_Spike_1
    Sample	Wells	Value	R	Result	MeanResult	CVs	CvResult	%SpikeRecovery	
    
    Group: Matrix_Spike_2
    Sample	Well	Value	R	Result	MeanResult	CVs	CVResult	%SpikeRecovery	
    
    Group: Standards
    Sample	Concentration ng/mL	Wells	Value	MeanValue	SD	CVs	CV_Result
    
    Group: Samples
    Sample	Wells	Value	R	Result	MeanResult	SD	CVs	CVresult

I believe this requires adding support for `REDUCED` files, so that's
what this PR is for.

I think we also need the "Acceptance Criteria" data from the `Note'
section, in which case I'll be following up with a larger PR to add some
functionality there.

---------

Co-authored-by: Nathan Stender <nathan.stender@benchling.com>