feat: JSON-to-CSV - improve list parsing to better support datacubes (#1124)
Currently, when two elements are not in the same root, we always cross
the parsed datasets.

ASM datacubes store the x/y dimensions in lists in separate paths, which
means they get crossed, so that for:

```
data : {
  dimensions: [[1,2]],
  measures: [[0.25, 0.5] 
}
```
you get
```
1,0.25
1,0.5
2,0.25
2,0.5
```

Now, we detect if all parsed datasets being combined have the same
length, and if so, just concat them.

This gets us the expected result:
```
1,0.25
2,0.5
``` 
for datacube extraction, while still crossing with downstream nested
values